pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lunchemail"
version = "0.1.0"
authors = [{name="Brian Holland", email="nosed-plosive.0n@icloud.com"}]
description = "Send out email for Friday's lunch"
readme = "README.md"
dependencies = [
    "appdirs==1.4.4",
    "ruamel.yaml==0.17.28",
    "ruamel.yaml.clib==0.2.7",
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    "black==23.3.0",
    "mypy==1.3.0",
    "ruff==0.0.270",
]

[project.scripts]
lunchemail = "lunchemail:main"

[project.urls]
"Homepage" = "https://github.com/brianholland99/lunchemail"
"Bug Tracker" = "https://github.com/brianholland99/lunchemail/issues"

[tool.black]
# TODO: Adjust line-length as desired.
line-length = 79
target-version = ["py38", "py39", "py310", "py311"]
